Fuse the decode: 574 ms -> 23 ms, 38 min -> 1.5 min on a 355B body
Four profiler-directed steps.  The scan was not a scan (ConvCode.step is a
shift register, so the state at row r is the previous 'memory' select bits and
the replay is a windowed function with O(1) depth).  The bit window is built by
doubling, not by a 'memory'-deep lagged-OR ladder.  The 128-entry subset and
transition tables are cached instead of rebuilt per call.  And every plane and
intermediate was int64 -- eight bytes per three bits -- across ~15 unfused
passes; BODY and the decoder's output are now uint8 and replay+completion is
one torch.compile'd kernel.

This removes the reason for a materialised sidecar: at 1.5 min for a 355B body,
decoding at load is cheaper than the disk a second copy would cost.  It does
not change VRAM, which stays 4.5 bpp -- only an in-kernel decoder would.

Two dtype landmines exposed and fixed: a uint8 index tensor is a boolean mask
in torch, so COMPLETION is widened at the reader while BODY stays narrow; and
decode_codes vs decode_codes_mixed had disagreed about the dtype of a nibble.

fused == eager bit-for-bit over memory in {3,4,5,6,8} x R in {1,2,3}; 195 tests
green with fusion on and under TESSERA_FUSED_REPLAY=0.

+++ src/tessera/encode.py
@@ -54,7 +54,7 @@ class EncodedUnit:
     rates: "tuple[int, ...]"
     anchors: torch.Tensor        # [rows, cols] int64, index into forest anchors
     codes: torch.Tensor          # [rows, cols] int64, E2M1 nibble after Stage C
-    body_bits: torch.Tensor      # [rows, cols] int64, the R input bits per position
+    body_bits: torch.Tensor      # [rows, cols] uint8, the R input bits per position
     completion_bits: torch.Tensor  # [rows, cols] int64, the c completion bits
     scale_base: torch.Tensor     # [groups] uint8, E8M0 exponent byte
     scale_refine: torch.Tensor   # [halves] uint8, 4-bit refinement word
@@ -254,7 +254,7 @@ def encode_unit(
     targets = work / scale
 
     anchors = torch.zeros(rows, cols, dtype=torch.long, device=device)
-    body_bits = torch.zeros(rows, cols, dtype=torch.long, device=device)
+    body_bits = torch.zeros(rows, cols, dtype=torch.uint8, device=device)
     completion_bits = torch.zeros(rows, cols, dtype=torch.long, device=device)
     codes = torch.zeros(rows, cols, dtype=torch.long, device=device)
     values = e2m1_value_table(device)
@@ -276,7 +276,7 @@ def encode_unit(
         per_pos = values[reachable][a]
         c_bits = ((sub.unsqueeze(2) - per_pos) ** 2).argmin(dim=2)
         anchors[:, which] = a
-        body_bits[:, which] = b
+        body_bits[:, which] = b.to(torch.uint8)
         completion_bits[:, which] = c_bits
         codes[:, which] = reachable[a, c_bits]
